About our Innovative Project
The proposed system is developed to deal with the home automation and industrial applications that are required for a smart city and its infrastructure. This project involves the capacity to use internet connected devices that may operate households and industrial appliances. It might involve preprogramming for smart lighting, intelligent heating and cooling and automated windows and doors all of which are connected via a central hub and can be operated remotely via smart phones.